Fixed switch in board.check, first move now works!
Missing break statements were causing the switch statement not to work,
which murdered everything else. Fixing that and tweaking checkline to
use a copy of the piece rather than a pointer to it means the first move
now works properly. Huzzah! black can't move and his placed pieces don't
show up though, working on that now.